downloader/aria2: split monolithic adapter into focused files (rpc, ops, files, delete, events); no behavior change
